Calculator events working
